refactor partials translations
in the diff output lines that had been translated due to being a partial
the translation was being used rather than the original. This commit
creates a means of looking up translations, and is the first step in
fixing this issue. The intention is then that we can do a lookup on a
change and see if there's a translation to replace it with.