fix _group_into_blocks method
this method was not appending the last line of the config to the groups
list and so it was being missed out of the config. Usually though the
last line of a config is 'end' so it hadn't mattered, and this line had
not been included in the test data and so had gone unnoticed. Would be
good to include some smaller test data sets as well to be more
manageable.