re-write variables functionality
I think, maybe, this time, despite so often being so close, yet so, so
far, with a soundtrack of screaming kids and tiredness, I have good
enough for the time being code that enables variables in the baseline.
Though I'm sure there will be a corner-case that undermines all this,
would be advisable to start documenting any rules for this. Also, does
not yet allow changing the delimiters.