implement binary search method.:+1:
Now instead of horrible nested for loops we have ALGORITHMS! :smiley:
Using a hash for lines that don't have any variables we can do a fast
lookup, and for those that do contain variables we can do a binary
search which is more elegant and speeds things up as well. Super.