Don't have a default ignores file
I think it's going to be easier all round if we don't specify a default
ignores file. Firstly this means specifying an ignores file is a
proactive thing and there can be no confusion over whether an ignores
file is being used. Secondly it means we don't have to specify a
particular name for it, and users can call it what they want.